Fixed bug for underflow subtract of `usize`

+++ src/bot/apply.rs
@@ -314,7 +314,7 @@ pub(super) async fn apply_view(
         }
         Some(option) => {
             if option == "PREV" {
-                handle_re_show_options(&bot, &dialogue, &q.from.username, applications, prefix, max(0, start-8), 8, msg_id).await?;
+                handle_re_show_options(&bot, &dialogue, &q.from.username, applications, prefix, max(0, start as i64 - 8) as usize, 8, msg_id).await?;
             } else if option == "NEXT" {
                 let entries_len = applications.len();
                 handle_re_show_options(&bot, &dialogue, &q.from.username, applications, prefix, if start+8 < entries_len { start+8 } else { start }, 8, msg_id).await?;
